+++ scripts/cameraRawInstaller.sh
@@ -55,3 +55,4 @@ function install_cameraRaw() {
 
 main
 
+