+++ scripts/launcher.sh
@@ -121,3 +121,4 @@ fi
 
 exit $exit_code
 
+