+++ scripts/sharedFuncs.sh
@@ -410,3 +410,4 @@ function load_paths() {
     unset datafile
 }
 
+