+++ scripts/uninstaller.sh
@@ -100,3 +100,4 @@ function ask_question() {
 load_paths "true"
 main
 
+