+++ troubleshoot.sh
@@ -380,3 +380,4 @@ echo "https://github.com/Gictorbit/photoshopCClinux/issues"
 echo ""
 
 
+